refactor(page settings): 設定ページのロジックをstoreに移行した
